code/image-classification/requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.123.0
flatbuffers==25.9.23
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
keras==3.12.0
libclang==18.1.1
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.4
namex==0.1.0
numpy==2.3.5
opt-einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.1
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pytest==9.0.1
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
